requirements.txt
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
matplotlib>=3.4.0
morse-translator==0.1
pyzbar>=0.1.8
pydub>=0.25.1
scipy>=1.7.0
scikit-image>=0.18.0
python-dotenv>=0.19.0
qrcode>=7.3.0
python-barcode>=0.13.1
pytesseract>=0.3.8
librosa>=0.8.1
cryptography>=41.0.0
pyttsx3>=2.90
braille-translator>=0.1.0
